pyproject.toml
[tool.poetry]
name = "message-sender"
version = "0.1.0"
description = ""
authors = ["Devlin Duldulao <webmasterdevlin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
azure-servicebus = "^7.12.2"
asyncio = "^3.4.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
